move calccrow() & torad() to src/common/helperfunction.js
